+++ tests/spec/directives/timespan-picker.spec.js
@@ -3,7 +3,7 @@ describe('luidTimespan', function(){
 	beforeEach(module('underscore'));
 	beforeEach(module('ngMock'));
 	beforeEach(module('lui'));
-	beforeEach(module('lui.directives'));
+	// beforeEach(module('lui.directives'));
 
 	var moment, $scope, isolateScope, $compile, elt, input;
 	beforeEach(inject(function (_moment_, _$rootScope_, _$compile_) {