+++ ts/iban/iban.controller.ts
@@ -1,17 +1,20 @@
-module Lui.Directives {
+module lui.iban {
 	"use strict";
 
 	export class LuidIbanController {
 		public static IID: string = "luidIbanController";
-		public static $inject: Array<string> = ["$scope"];
+		public static $inject: Array<string> = ["$scope", "iban"];
 		private $scope: ILuidIbanScope;
 		private ngModelCtrl: ng.INgModelController;
 		private countryInput: ng.IAugmentedJQuery;
 		private controlInput: ng.IAugmentedJQuery;
 		private bbanInput: ng.IAugmentedJQuery;
 
-		constructor($scope: ILuidIbanScope) {
+		private ibanChecker: IBANStatic;
+
+		constructor($scope: ILuidIbanScope, iban: IBANStatic) {
 			this.$scope = $scope;
+			this.ibanChecker = iban;
 			this.initScope();
 		}
 
@@ -31,7 +34,7 @@ module Lui.Directives {
 			};
 			(<ILuidIbanValidators>this.ngModelCtrl.$validators).iban = (): boolean => {
 				if (!!this.ngModelCtrl.$viewValue) {
-					return IBAN.isValid(ngModelCtrl.$viewValue);
+					return this.ibanChecker.isValid(ngModelCtrl.$viewValue);
 				}
 				return true;
 			};
@@ -102,5 +105,5 @@ module Lui.Directives {
 		}
 	}
 
-	angular.module("lui.directives").controller(LuidIbanController.IID, LuidIbanController);
+	angular.module("lui.iban").controller(LuidIbanController.IID, LuidIbanController);
 }